requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.5
djangorestframework==3.13.1
filelock==3.7.1
idna==3.3
lxml==4.9.0
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2==2.9.3
pyTelegramBotAPI==4.5.1
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7